flask/requirements.txt
Flask==3.0.3
gunicorn==21.2.0
lxml_html_clean==0.1.1
newspaper3k==0.2.8
openai==1.23.2
pyppeteer==2.0.0
python-dotenv==1.0.1
readability-lxml==0.8.1
